requirements.txt
Flask==2.1.0
gunicorn==20.1.0
google-cloud-language>=2.4.1